Week 6 - Retrieval Augmented Generation/requirements.txt
streamlit
openai
qdrant-client
sentence-transformers
pandas
python-docx
pymupdf
beautifulsoup4
lxml